Refactor galaxy_install.datatypes into galaxy.
