+++ lib/galaxy/tool_shed/galaxy_install/datatypes/custom_datatype_manager.py
