+++ lib/tool_shed/galaxy_install/installed_repository_manager.py
@@ -9,9 +9,9 @@ import shutil
 from sqlalchemy import and_, false, true
 
 from galaxy import util
+from galaxy.tool_shed.galaxy_install.datatypes import custom_datatype_manager
 from galaxy.tool_shed.galaxy_install.tools import data_manager
 from galaxy.tool_shed.galaxy_install.tools import tool_panel_manager
-from tool_shed.galaxy_install.datatypes import custom_datatype_manager
 from tool_shed.galaxy_install.metadata.installed_repository_metadata_manager import InstalledRepositoryMetadataManager
 from tool_shed.galaxy_install.repository_dependencies import repository_dependency_manager
 from tool_shed.util import common_util
