+++ lib/tool_shed/galaxy_install/tool_migration_manager.py
@@ -11,12 +11,12 @@ import threading
 from collections import OrderedDict
 
 from galaxy import util
+from galaxy.tool_shed.galaxy_install.datatypes import custom_datatype_manager
 from galaxy.tool_shed.galaxy_install.tools import tool_panel_manager
 from galaxy.tool_shed.tools import data_table_manager
 from galaxy.tools.toolbox import ToolSection
 from galaxy.tools.toolbox.parser import ensure_tool_conf_item
 from tool_shed.galaxy_install import install_manager
-from tool_shed.galaxy_install.datatypes import custom_datatype_manager
 from tool_shed.galaxy_install.metadata.installed_repository_metadata_manager import InstalledRepositoryMetadataManager
 from tool_shed.util import (
     basic_util,
